+++ core/types/snapshot.go
@@ -90,6 +90,7 @@ const (
 	SettlementSnapshot             SnapshotNamespace = "settlement"
 	HoldingAccountTrackerSnapshot  SnapshotNamespace = "holdingAccountTracker"
 	EthereumOracleVerifierSnapshot SnapshotNamespace = "ethereumoracleverifier"
+	TeamsSnapshot                  SnapshotNamespace = "teams"
 
 	MaxChunkSize   = 16 * 1000 * 1000 // technically 16 * 1024 * 1024, but you know
 	IdealChunkSize = 10 * 1000 * 1000 // aim for 10MB