+++ CHANGELOG.md
@@ -47,6 +47,7 @@
 - [8770](https://github.com/vegaprotocol/vega/issues/8770) - Add `PERPS` to integration tests.
 - [8763](https://github.com/vegaprotocol/vega/issues/8763) - Periodic settlement data endpoint.
 - [8920](https://github.com/vegaprotocol/vega/issues/8920) - Emit events when something happens to teams.
+- [8917](https://github.com/vegaprotocol/vega/issues/8917) - Support teams engine snapshots
 
 ### 🐛 Fixes